requirements.txt
numpy
scipy
scikit-learn
tqdm
pandas
pyarrow
PyYAML
litellm
pydantic>=2
tenacity
transformers>=4.57.0,<5
accelerate>=0.26.0
gliclass>=0.1.16
